feat: Add custom exception classes for user-related errors

+++ src/main/java/ifsul/agileproject/rachadinha/service/impl/UserServiceImpl.java
@@ -2,11 +2,12 @@ package ifsul.agileproject.rachadinha.service.impl;
 
 import ifsul.agileproject.rachadinha.domain.dto.UserDTO;
 import ifsul.agileproject.rachadinha.domain.entity.User;
-import ifsul.agileproject.rachadinha.exceptions.EmailAlreadyUsedException;
+import ifsul.agileproject.rachadinha.exceptions.*;
 import ifsul.agileproject.rachadinha.mapper.UserMapper;
 import ifsul.agileproject.rachadinha.repository.UserRepository;
 import ifsul.agileproject.rachadinha.service.UserService;
 import lombok.AllArgsConstructor;
+
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -24,7 +25,7 @@ public class UserServiceImpl implements UserService {
   @Override
   public User saveUser(UserDTO userDTO) {
     if (userRepository.existsByEmail(userDTO.getEmail())) {
-      throw new EmailAlreadyUsedException("Endereco de email ja utilizado");
+      throw new EmailAlreadyUsedException(userDTO.getEmail());
     }
     User user = userMapper.apply(userDTO);
     return userRepository.save(user);
@@ -32,11 +33,17 @@ public class UserServiceImpl implements UserService {
 
   @Override
   public Optional<User> findUserById(Long id) {
+    if (!userRepository.existsById(id)) {
+      throw new UserNotFoundException(id);
+    }
     return userRepository.findById(id);
   }
 
   @Override
   public void deleteUserById(Long id) {
+    if (!userRepository.existsById(id)) {
+      throw new UserNotFoundException(id);
+    }
     userRepository.deleteById(id);
   }
 
@@ -47,7 +54,14 @@ public class UserServiceImpl implements UserService {
 
   @Override
   public User login(String email, String password) {
-    return userRepository.findByEmailAndPassword(email, password);
+    if (!userRepository.existsByEmail(email)) {
+      throw new UserNotFoundException(email);
+    } 
+    User foundUser = userRepository.findByEmailAndPassword(email, password);
+    if (foundUser == null) {
+      throw new IncorrectUserPasswordException(email);
+    }
+    return foundUser;
   }
 
   @Override
@@ -61,16 +75,17 @@ public class UserServiceImpl implements UserService {
 
     Predicate<String> isNotEmpty = (value) -> !value.isEmpty();
 
-    if (user.isPresent()) {
-      User updatedUser = user.get();
-
-      Optional.ofNullable(userDTO.getName()).filter(isNotEmpty).ifPresent(updatedUser::setName);
-      Optional.ofNullable(userDTO.getEmail()).filter(isNotEmpty).ifPresent(updatedUser::setEmail);
-      Optional.ofNullable(userDTO.getPassword()).filter(isNotEmpty).ifPresent(updatedUser::setPassword);
-
-      return userRepository.save(updatedUser);
+    if (!user.isPresent()) {
+      throw new UserNotFoundException(userDTO.getEmail());
     }
-    throw new RuntimeException();
+
+    User updatedUser = user.get();
+
+    Optional.ofNullable(userDTO.getName()).filter(isNotEmpty).ifPresent(updatedUser::setName);
+    Optional.ofNullable(userDTO.getEmail()).filter(isNotEmpty).ifPresent(updatedUser::setEmail);
+    Optional.ofNullable(userDTO.getPassword()).filter(isNotEmpty).ifPresent(updatedUser::setPassword);
+
+    return userRepository.save(updatedUser);
   }
 
   @Override
@@ -79,20 +94,21 @@ public class UserServiceImpl implements UserService {
   }
 
   @Override
-  public User resetPassword(Long userId, String oriPass, String newPass) throws Exception {
+  public User resetPassword(Long userId, String oriPass, String newPass) {
     Optional<User> user = userRepository.findById(userId);
 
-    if(user.isPresent()){
-      User usuario = user.get();
-
-      if(usuario.getPassword().equals(oriPass)){
-        usuario.setPassword(newPass);
-        return userRepository.save(usuario);
-      } else{
-        throw new Exception("Senha errada");
-      }
-    } else{
-      throw new Exception("Usuário não existe");
+    if (!user.isPresent()) {
+      throw new UserNotFoundException(userId);
     }
+
+    User usuario = user.get();
+
+    if(!usuario.getPassword().equals(oriPass)){
+      throw new IncorrectUserPasswordException(usuario.getEmail());
+    }
+
+    usuario.setPassword(newPass);
+
+    return userRepository.save(usuario);
   }
 }
